+++ src/duplocloud/argtype.py
@@ -1,4 +1,6 @@
 from typing import NewType
+import argparse
+import yaml
 
 class Arg(NewType):
   """Command Argument Type
@@ -60,3 +62,10 @@ class Arg(NewType):
   @property
   def positional(self):
     return len(self._flags) == 0
+
+class YamlAction(argparse.Action):
+  def __init__(self, option_strings, dest, nargs=None, **kwargs):
+    super().__init__(option_strings, dest, **kwargs)
+  def __call__(self, parser, namespace, value, option_string=None):
+    data = yaml.load(value, Loader=yaml.FullLoader)
+    setattr(namespace, self.dest, data)
